Wrap all the errors into an MQTT message, not just the first one
When there is a non fatal error in the init step of the c8y mapper,
one needs to proceed, i.e. to send an error for the missed step,
then to continue with the other init messages.

Before that fix, only parts of the init messages were sent after the
first initialization error.

That said the root issue that were breaking the tests is that the
path to the operations directory is hardcoded and non-existing in the
test environment. See #814 for a fix of the root cause.

Signed-off-by: Didier Wenzek <didier.wenzek@free.fr>